@@ -6,67 +6,70 @@
     "ng": "ng",
     "prestart": "ts-node git.version.ts",
     "prebuild": "ts-node git.version.ts",
-    "start": "ng serve",
-    "start:timeseries": "ng serve timeseries",
+    "postinstall": "ts-node scripts/adjust-install.ts",
+    "start": "ng serve timeseries",
     "start:complete": "ng serve complete",
     "prod": "ng serve --prod",
-    "build": "ng build --prod",
-    "build:timeseries": "ng build timeseries --prod",
-    "prebuild:timeseries": "npm run prebuild",
+    "build": "ng build timeseries --prod",
     "build:complete": "ng build complete --prod",
     "prebuild:complete": "npm run prebuild",
     "test": "ng test",
     "lint": "ng lint --type-check",
     "e2e": "ng e2e",
-    "build-netlify": "npm run build && cp ./netlify/* ./dist/complete -r",
-    "i18n-extract": "ngx-translate-extract -i ./src -o ./src/assets/i18n/*.json --clean --sort --format namespaced-json"
+    "build-netlify": "npm run build:complete && cp ./netlify/* ./dist/complete -r",
+    "i18n-extract": "ngx-translate-extract -i ./src -o ./src/assets/i18n/*.json --clean --sort --format namespaced-json --marker marker",
+    "bundle-war": "./node_modules/ts-node/dist/bin.js scripts/bundle-war.ts",
+    "ng-high-memory": "node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.2",
-    "@angular/common": "^6.1.2",
-    "@angular/compiler": "^6.1.2",
-    "@angular/core": "^6.1.2",
-    "@angular/forms": "^6.1.2",
-    "@angular/http": "^6.1.2",
-    "@angular/platform-browser": "^6.1.2",
-    "@angular/platform-browser-dynamic": "^6.1.2",
-    "@angular/router": "^6.1.2",
-    "@helgoland/caching": "0.0.1-beta.2",
-    "@helgoland/control": "0.0.1-beta.2",
-    "@helgoland/core": "0.0.1-beta.2",
-    "@helgoland/d3": "0.0.1-beta.2",
-    "@helgoland/depiction": "0.0.1-beta.2",
-    "@helgoland/favorite": "0.0.1-beta.2",
-    "@helgoland/map": "0.0.1-beta.2",
-    "@helgoland/modification": "0.0.1-beta.2",
-    "@helgoland/permalink": "0.0.1-beta.2",
-    "@helgoland/plotly": "0.0.1-beta.2",
-    "@helgoland/selector": "0.0.1-beta.2",
-    "@helgoland/time": "0.0.1-beta.2",
-    "@helgoland/time-range-slider": "0.0.1-beta.2",
+    "@angular/animations": "^8.2.4",
+    "@angular/common": "^8.2.4",
+    "@angular/compiler": "^8.2.4",
+    "@angular/core": "^8.2.4",
+    "@angular/forms": "^8.2.4",
+    "@angular/platform-browser": "^8.2.4",
+    "@angular/platform-browser-dynamic": "^8.2.4",
+    "@angular/router": "^8.2.4",
+    "@helgoland/caching": "1.0.0-beta.4",
+    "@helgoland/control": "1.0.0-beta.4",
+    "@helgoland/core": "1.0.0-beta.4",
+    "@helgoland/d3": "1.0.0-beta.4",
+    "@helgoland/depiction": "1.0.0-beta.4",
+    "@helgoland/favorite": "1.0.0-beta.4",
+    "@helgoland/map": "1.0.0-beta.4",
+    "@helgoland/modification": "1.0.0-beta.4",
+    "@helgoland/permalink": "1.0.0-beta.4",
+    "@helgoland/plotly": "1.0.0-beta.4",
+    "@helgoland/selector": "1.0.0-beta.4",
+    "@helgoland/time": "1.0.0-beta.4",
+    "@helgoland/time-range-slider": "1.0.0-beta.4",
     "@ng-bootstrap/ng-bootstrap": "^4.2.1",
     "@ngx-translate/http-loader": "^2.0.1",
     "@types/d3": "^5.7.2",
+    "archiver": "^3.1.1",
     "bootstrap": "^4.3.1",
     "bootstrap-slider": "^10.6.1",
     "core-js": "^2.5.7",
     "font-awesome": "^4.7.0",
+    "fs": "0.0.1-security",
     "leaflet-geosearch": "^2.7.0",
     "rxjs": "^6.5.2",
     "rxjs-compat": "^6.5.2",
-    "zone.js": "^0.8.26"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.7.3",
-    "@angular/cli": "^6.1.3",
-    "@angular/compiler-cli": "^6.1.2",
-    "@angular/language-service": "^6.1.2",
-    "@biesbjerg/ngx-translate-extract": "^2.3.4",
+    "@angular-devkit/build-angular": "~0.803.2",
+    "@angular/cli": "^8.3.2",
+    "@angular/compiler-cli": "^8.2.4",
+    "@angular/language-service": "^8.2.4",
+    "@biesbjerg/ngx-translate-extract": "^3.0.4",
+    "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
     "@types/jasmine": "^2.8.8",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "^10.12.18",
-    "codelyzer": "^4.4.3",
+    "codelyzer": "^5.0.1",
     "file-loader": "^1.1.6",
     "har-validator": "^5.1.3",
     "jasmine-core": "^3.4.0",
@@ -82,6 +85,7 @@
     "qs": "^6.7.0",
     "ts-node": "^7.0.1",
     "tslint": "^5.18.0",
-    "typescript": "2.9.2"
+    "typescript": "3.5.3",
+    "webpack-cli": "^3.3.7"
   }
 }
