@@ -5,9 +5,13 @@ import {
     DatasetOptions,
     LocalStorage,
     RenderingHintsDatasetService,
+    SettingsService,
     Time,
     Timespan,
 } from '@helgoland/core';
+import { duration, MomentInputObject } from 'moment';
+
+import { HelgolandSettings } from './../../settings/helgoland-settings';
 
 const TIMESERIES_OPTIONS_CACHE_PARAM = 'timeseriesOptions';
 const TIMESERIES_IDS_CACHE_PARAM = 'timeseriesIds';
@@ -22,6 +26,7 @@ export class TimeseriesService extends RenderingHintsDatasetService<DatasetOptio
         protected localStorage: LocalStorage,
         protected timeSrvc: Time,
         protected api: DatasetApiInterface,
+        private settingsSrvc: SettingsService<HelgolandSettings>,
         private color: ColorService
     ) {
         super(api);
@@ -31,6 +36,7 @@ export class TimeseriesService extends RenderingHintsDatasetService<DatasetOptio
     protected createStyles(internalId: string) {
         const options = new DatasetOptions(internalId, this.color.getColor());
         options.generalize = false;
+        options.lineWidth = 2;
         return options;
     }
 
@@ -49,7 +55,32 @@ export class TimeseriesService extends RenderingHintsDatasetService<DatasetOptio
         const options = this.localStorage.loadArray<DatasetOptions>(TIMESERIES_OPTIONS_CACHE_PARAM);
         if (options) { options.forEach(e => this.datasetOptions.set(e.internalId, e)); }
         this.datasetIds = this.localStorage.loadArray<string>(TIMESERIES_IDS_CACHE_PARAM) || [];
-        this.timespan = this.timeSrvc.loadTimespan(TIME_CACHE_PARAM) || this.timeSrvc.initTimespan();
+        this.timespan = this.timeSrvc.loadTimespan(TIME_CACHE_PARAM) || this.initTimespan();
+    }
+
+    private initTimespan(): Timespan {
+        const defDuration = this.settingsSrvc.getSettings().defaultTimeseriesTimeduration;
+        if (defDuration && defDuration.duration && defDuration.align) {
+            return this.generateTimespan(defDuration.duration, defDuration.align);
+        } else {
+            return this.timeSrvc.initTimespan();
+        }
+    }
+
+    // TODO: use time service for this
+    private generateTimespan(defaultTimeseriesTimeduration: MomentInputObject, align: 'start' | 'center' | 'end'): Timespan {
+        const now = new Date();
+        const d = duration(defaultTimeseriesTimeduration);
+        switch (align) {
+            case 'start':
+                return new Timespan(now.getTime(), now.getTime() + d.asMilliseconds());
+            case 'end':
+                return new Timespan(now.getTime() - d.asMilliseconds(), now.getTime());
+            case 'center':
+            default:
+                const half = d.asMilliseconds() / 2;
+                return new Timespan(now.getTime() - half, now.getTime() + half);
+        }
     }
 
 }
